+++ src/main/java/cn/edu/buaa/scholarshipserver/dao/DataScholarDao.java
@@ -9,7 +9,7 @@ import java.util.List;
 public interface DataScholarDao extends ElasticsearchRepository<DataScholar,Long> {
     public DataScholar findByAuthorId(Long authorId);
     public List<DataScholar> findByNormalizedName(String normalizedName);
-    public List<DataScholar> findByScholarId(Long scholarId);
+    public List<DataScholar> findByScholarId(Integer scholarId);
 
 }
