+++ src/main/java/cn/edu/buaa/scholarshipserver/dao/PatentScholarDao.java
@@ -8,5 +8,5 @@ import org.springframework.data.elasticsearch.repository.ElasticsearchRepository
 import java.util.List;
 
 public interface PatentScholarDao extends ElasticsearchRepository<Patent_Scholar,Long> {
-        public List<Patent_Scholar> findByScholarId(Long scholarId);
+        public List<Patent_Scholar> findByScholarId(Integer scholarId);
 }
